Set Mesos agent executor_environment_variables
This commit fixes the Docker task failures on the DCOS Windows nodes.

The Mesos slave doesn't inherit the system environment variables.
Thus, the system PATH is not visible to the Mesos slave and when
trying to spawn a Docker task via Mesos, the Docker binary is not
found in the PATH.

We set a sane path for the Mesos slave via the
'--executor_environment_variables' parameter.