Set hostname explicitly for the Mesos slave
All the Mesos Windows slaves report the same hostname which is
resolvable to the IP address of one of the agent. This is wrong
as they all have different hostnames.

For each slave, we set the hostname to the IP address of node,
just as it's being done with the Linux slaves.