Change the default service behaviour creation
Until now we restarted the service every time it did not gracefully stopped,
with 1 second delay between restarts.

This patch proposes a new default behaviour.

The first two restart come instantly while for the third restart we add a 30
second timeout.

We reset the flags above if and only if the service was running for more than
40 seconds.

I.E.:
In the case a service is broken now it will have the following behaviour
from the point of view of SCM:
```(start *)``` - restart (0 ms) - restart (0 ms) - restart (30 seconds) - restart (30 seconds) - ...
We will go back to ```(start *)``` if the service was up and running for
>= 40 seconds.

This will aid debugging and cause less spam in logs.

Signed-off-by: Alin Gabriel Serdean <aserdean@ovn.org>