Add executor registration timeout to the mesos agent on Windows
Windows docker images have a great footprint and will take some time until
they are pulled to the local image registry.

Mesos agent has a default 1 min timeout for the registration, unfortunately
this is not enough for the size of the images.

Bump the timeout so the tasks won't be removed if they are in pulling state.

Signed-off-by: Alin Gabriel Serdean <aserdean@ovn.org>