Add DCOSWindowsAgentSetup.ps1
This is meant to be a standalone script called directly by the
custom data.

Its main purpose is the to fetch the Windows agent blob files,
and to execute `AGENT_BLOB_DIR\scripts\agent-setup.ps1`.

The `agent-setup.ps1` is the PowerShell script in the blob
that will set-up the DC/OS Windows agent with the binaries
from there.