Add stop signal SIGKILL functionality to units
Until now we try to send a CTRL_C event to the process and wait for them
to gracefully shutdown.

With this patch we check if the signal was SIGKILL and save the state.
In the case it was specified we just terminate the process tree.

This patch also fixes some handle leaks that were found while testing.

Signed-off-by: Alin Gabriel Serdean <aserdean@ovn.org>